Refactored repository structure: moved TaskRepository, TaskRepositoryCustom, and TaskRepositoryImpl to repository/task/
